WIP Rename language modeling module
