+++ fish_speech_core/lib/models/lm/generate/mod.rs
@@ -1,6 +1,6 @@
-pub mod batch;
 pub mod single_batch;
+pub mod static_batch;
 mod utils;
 
-pub use batch::{generate_static_batch, BatchGenerator};
 pub use single_batch::{generate_blocking, generate_blocking_with_hidden, SingleBatchGenerator};
+pub use static_batch::{generate_static_batch, BatchGenerator};
