+++ fish_speech_core/lib/models/lm/utils/mod.rs
