+++ fish_speech_core/lib/models/mod.rs
@@ -1,2 +1,2 @@
+pub mod lm;
 pub mod vqgan;
-pub mod text2semantic;
