+++ server/lib/lib.rs
@@ -3,8 +3,8 @@ pub mod handlers;
 pub mod state;
 
 use candle_core::{Device, Tensor};
-use fish_speech_core::models::text2semantic::utils::encode::PromptEncoder;
-use fish_speech_core::models::text2semantic::utils::sample::load_prompt_text;
+use fish_speech_core::models::lm::utils::encode::PromptEncoder;
+use fish_speech_core::models::lm::utils::sample::load_prompt_text;
 use fish_speech_core::models::vqgan::config::WhichLM;
 use serde::Deserialize;
 use std::collections::HashMap;
